Ensure that the default namespace root is authorized.
By changing the default value of the --v23.namespace.root flag to
include the blessings expected from it, we aim to close the loop
on secure communication: Now, by default, all command-line programs
will authenticate the root mounttable as well. If the root mounttable
presents a blessing that is unexpected, the program will simply not
communicate with it - not reveal its public key, its blessings etc.

Closes veyron/release-issues#98

Change-Id: Icead2dc1251bead405030700d984e55cb9b6c716