ref: Introduce client.Call that retries application level errors.
If a server returns a retryable verror, client.Call will retry the
call.
client.Call is a synchronous substitute for client.StartCall/call.Finish.

I had to add the method to the Client interface instead of directly in the
v23/rpc repository so that the implementation can be in x/ref/profile/...

MultiPart: 2/2

Change-Id: I90c849070b71ded937406b60d59344d866e92c63