wsprd: Added the ability to pass Blessings in RPCs at the top level.
This does not work if the blessings are nested in another data
structure.  This is a short-term hack to get all this to work.  The
long term solution is to bite the bullet and pass the whole blessing
object back and forth between JS and Go with aggressive caching on
both ends.

This change also allows for grabbing the default blessings for the
principal of the runtime

MultiPart: 1/2
Change-Id: I0aff94add4056c76612ac2a29412d565ec200be9