services/device/internal/impl: Use suidhelper kill functionality
Use killing via suidhelper to enforce app exit after a Stop() call.
While doing so, enable the deadline parameter to the Stop() call.

Also, use the suidhelper killing to kill off any newly-started app
that either fails to handshake properly or appears to lie about its pid
during the handshake.

In order to get this done without passing the suidhelper path around
everywhere, refactored the way that suidhelper is called to
centralize the knowledge of the suidhelper path and its flags in
helper_manager.go.

Change-Id: I84adcf874814bb040cfe8668da65e0193246c704