TBR: rpc/stream/vc: Fix race in vc_test that is currently affecting Jenkins.
This race was introduced in cl 9932.  In the process of that change I
made HanshakeAcceptedVC blocking mistakenly which led me to introduce
more concurrency into the test.  The blocking problem was caught and
fixed during review, but the additional concurrency in the test remained.
This can cause a race because, in the test, the client can start sending
data to the server VC before it's listener has been created which
eventually causes a panic.

Change-Id: I3a5646ab60db5edf5e97afa85faa208804b6d822