TBR wspr: Fix wspr hanging when trying to unmount.
TBRing for suharsh so we can test this

PROBLEM: When wspr shuts down, server never stop.

REASON:
In WSPR, a customCaveatValiditor is used. This validator asks JS to
validate all caveats (except tpcavs). When a tab is closed, the JS
code disappears, so when the server tries to unmount, it is never
able to do so because its caveats never get validated.

SOLUTION:
The customCaveatValidator in wspr validates caveats in Go if the server
has been shutdown.

Change-Id: I14d907e6a5005b4df07d2f829d7cb7ef3f815900