ref: Change context mutators from SetFoo to WithFoo.
The motivation for this change is consistency, so our API becomes
easier to use.

Before this CL we were using the WithFoo style in the context
package, but the SetFoo style elsewhere, sometimes with an extra
"Context" thrown in for kicks.  After this CL we use WithFoo (and
GetFoo) everywhere.

Change-Id: Idc6959e4ea76ca6bd36501f4002145ffa8835c04
MultiPart: 2/4