services/identity: Force the OAuth approval prompt when seeking blessings.
Forcing this should allay spookiness fears for users that seekblessings
rarely (which they should) and thus forget about an earlier consent.

The listblessings page does not force this approval, but I think that is
fine since its prime purpose is to remind you when and how you used
seekblessings, so the forced approval would only add friction.

Closes veyron/release-issues#1943

Change-Id: I33fa283e2dfc0a18220966527599bb341ba980a8