rpc/client_test: Make test handle flaky error.
We check for "failed to encode arg" here because sometimes the server detects the
mismatched number of arguments, sends an error response, and closes the connection,
before the client gets through encoding the args. In this case the flow is closed and
encoding of args fails, preventing the client from calling call.Finish, and seeing
the error in the response. In the normal case network time dominates, so this case
will very rarely get hit, but since the client and server in this test are in the
same process we see this race quite a bit.

Addresses vanadium/issues#289

Change-Id: I2fb7648d272c132f3c3a17947bfaba0fd7012861