services/device/internal/starter: don't try publishing before claim
Most likely, the publish will fail since we don't have credentials, with the
effect that we hang for 30 seconds stuck in publish retry limbo.

This change rips out the local device mounttable pre-claim, and disables
publishing of the claimable service.  Instead, the endpoint of the claimable
service will be returned from starter.Start.

The device unit test is updated to reflect these changes -- in particular, we
manually mount the claimable service's endpoint into the global MT for discovery
by the test fixtures.  This ends up also simplifying the post-claim 'wait for
remount' logic.

The integration test is also updated -- we now grab the claimable endpoint from
the device manager's log.

Change-Id: I2426090be009e1f9c4b875aaf152c19bde8fa96a