services/proxy/proxyd: Implement access control at the proxy server.
This change makes proxyd support access control by allowing only servers
that present a blessing that matches AccessList to export themselves
through the proxy.

The change itself is backward compatible, if proxyd built with this
change is started with --access-list='{"In":["..."]}'. However,
my intention is to have this in before the release, so all servers
that use the proxy will already be using this new code.

I admit that I went for a bit of "speed of implementation" over elegance
in an attempt to get this in before we release publicly. I'm happy to do
cleanups after this is in.

MultiPart: 1/3

Closes vanadium/issues#33

Change-Id: Ic6413403f6bbca9c7b2bd330910a5c3e2adb0ef3