services/device/device: updateall: kill instance only if running; recover if err
Changing updateall to only attempt to kill an instance if the instance is found
to be running.  If we then try Kill and it returns an error, we still check if
it's running or not (since the device manager force-kills the app) and recover.
Before, we used to leave the app not running and exit, which not only left the
updates in a partial state, but also left previously running instances dead.

Also, update the kill deadline to 10 seconds, and use it consistently between
the 'kill' and 'updateall' commands.

Change-Id: Ib129b94536739b385401d68f509e860a786d7650