hello: Fix bug in error reporting.
I finally figured out what was going on.  If the mounttable fails
to produce the expected NAME variable before a timeout then it just returns
empty string.  We would then pass an empty string environment variable
for namespace which would cause the next binary to try to talk to the
production mounttable.  This made the test output very confusing.

Now I correctly report a fatal error if expected variables don't turn up.
I removed all the namespace related environment variable logic, since
using the commandline is safer and more transparent (it gets printed
in the output when the test fails).

Change-Id: I596d49b4c83fb37011ca8cc8cc86c5c6a2159ad7