wspr: Share an encoder and decoder for all js clients.
This improves RPC performance by 50% because the most expensive part
of making an rpc is encoding the type message.  This is really an
interim change, since the real solution is to have a separate type
flow that is shared by all messages from wspr to go.

This improves: https://github.com/vanadium/issues/issues/421

MultiPart: 2/3
Change-Id: Id9ebe77fbe04c143b2728af24003082d810d7ff4