"x/ref": SeekBlessings also sends public key of the principal tool
Currently the 'seekblessings' flow involves obtaining a macaroon
from the HTTPS identity service and sending that macaroon over a
Vanadium RPC in order to obtain a blessing. The returned blessing
is based on the email address encapsulated in the macaroon.

At the momement, the macaroon is a bearer credential and can be used
(to obtain a blessing) by anyone who holds it. This makes it an
attractive target for theft.

We plan to bind the macaroon to the public key of the principal
tool so that only the principal tool can use the macaroon to
obtain a blessing.

This would be accomplished in two steps:
1) The principal tool will be updated to also send its public
key in the HTTPS request to 'auth/google' endpoint of the Identity
Service.
2) The Identity service will be updated to include the public key
received from the tool in the issued macaroon, and only grant blessings
to clients whose public key matches the public key in the macaroon
presented by them

This CL carries out Step 1.

Change-Id: Ieb943abc5574a8a6f079b4c6676f42acf3f7f330