services/device/internal/impl: additional app liveness mechanism
Add another mechanism by which the device manager can determine if
apps are running when it restarts. This new default mechanism assumes
that the pid of process previously launched by the device manager that
does not respond to app cycle inquiries is the selfsame process and so
should remain in the device manger's purview.

Change-Id: I5a4dd5c5168bebe0e845c71448b32a45e74ff1bf