"x/ref": Never recognize a root for AllPrincipals
There is no good reason for a principal to recognize a root (public key)
for all blessing names. On the other hand, doing so makes the root
principal impersonate all other identity providers when talking to
the recognizing principal.

Therefore, this CL forbids BlessingRoots.Add from being call with
the AllPrincipals blessing pattern.

Change-Id: I750de9b7d86b640055509d50e85c1776171a8581