ref: Convert all cmdline users to cmdline2.
This also includes changes in cmdline2 to allow commands with
both Children and Runner set, as long as the runner doesn't take
any args.  This makes sense since the runner handles the no-arg
case, while the children handle the with-args case.  This was
already being used by "v23 buildcop" and "deviced", and it seemed
better to allow this behavior.

Also fixed a bug in the cmdline2 package where the "at least one
of Children and Runner" check wasn't returning an error, if the
bad Command wasn't the root.  Added tests for this case.

Finally, but also important, changed how global flags are
handled.  Previously I was hoping we could simply change
"v.io/x/ref/lib/flags" to check flag.Parsed and avoid
double-parsing the flags.  But the previous strategy always
merged flags into a new FlagSet, and then overwrote
flag.CommandLine with that value.  Our usage of ref/lib/flags
stashes the flag.CommandLine pointer away in an init function,
which breaks that strategy.

In addition, ref/lib/flags has some weird logic to set certain
flags to their "default values".  Presumably this was a hack to
deal with double-parsing the flags for flags that are specified
multiple times on the command line, and append their values.
This makes it questionable whether we can really depend on not
double-parsing flags.

I'll look into changing our flags so that they don't depend on
the multi-specification-append behavior, since that's just weird.
But I'll do that separately.

The actual conversions of all programs is mechanical.

MultiPart: 3/3

Change-Id: I3292256e5c8e4a6acf99c9b92c98b009a7fbfcf4