services/identity: Get rid of email classification.
The "email classifier" feature of identityd, which allowed it to
generate different blessings based on the email address of the requestor
is no longer a used feature, so remove it (and save some complexity).

Furthermore, for the identityd instance on https://dev.v.io/auth,
the plan is to have it start issuing shorter blessings of the form:
"dev.v.io/u/<email>" instead of the current
"dev.v.io/root/users/<email>". And in doing so, identityd will not
longer add the "users" prefix since it will be running as "dev.v.io/u".

This commit prepares identityd to work with both the new ("dev.v.io/u")
and old ("dev.v.io/root") blessings. Some code will be cleaned up
once the transition happens in production.

Closes vanadium/issues#468

Change-Id: I8c7dbe96f719e2e64b2315df13828e400fb94b7f