services/device/device: make tapes thread-safe, add unit tests for tapes
Make the tapes thread-safe s.t. they can be used for servicing rpc requests that
are performed in parallel.

Add unit test cases to cover tape functionality, and in particular, one test
case runs tape operations in parallel to verify thread-safety (in practice, the
test has been observed to fail when not using locking in tapes).

Also, change SetResponses to take a variable arg list instead of a slice.  This
is mostly for convenience, since the overwhelmingly common case is to only set
one response.

Change-Id: I90df6acfc02934fcffc0d042129cff530fcd52be