ref: Convert all programs in ref to use package cmdline.
The exceptions are inithelper and suidhelper, since they are
meant to use a minimal amount of code for easy auditing.

The rpc/benchmark/main is also left alone, since it relies on a
specific init sequence, and it's just for our own internal
benchmarking anyways.

Changed global flags to flags on the root command for many
programs, to get better help messages.

Change-Id: I48d61327862506ca50f9b8d57818a01941b256f1