cmd/principal: fulfill the spirit of v.io/i/201 by prominently printing expiry
The goal of https://github.com/veyron/release-issues/issues/1855 was to make it
obvious to users when their credentials were expired.  Suharsh's change
(v.io/c/11377) added the mechanism to blessings to expose this, but only used it
for principal dumpblessings (which is not something users typically run --
unless they already suspect that their blessings are expired).

This change adds the [EXPIRED] tag to the principal dump command (both with and
without the -s option). This will ensure that it shows up in the vbash prompt,
and also that people running principal dump by hand see it.

Change-Id: I384ae627817450016ab6a842c6658a718baf8ad3