ref: Add support for cmdline.Env.Vars
This adds envrionment variable support to cmdline.Env, and also
updates some programs to use this support.  The main benefit of
using cmdline.Env.Vars is that it makes testing easier, since you
can just pass in your own map.

We represent environment variables as map[string]string, which is
a nice simple form for lookups and mutations.  Originally I was
planning on using *envvar.Vars, but the delta-tracking support is
probably unnecessary, and the native Go map syntax is nicer.

A subsequent change will update the v23.Init and test.InitForTest
mechanisms to support explicit passing of args and envvars, and
will take advantage of this support.

MultiPart: 2/2
Change-Id: Ic9bb9a71b05fd73dafa630f069236c716ee33fb2