mounttablelib: developer brain in neutral - fixed.
When building the database from the config, I didn't copy permissions
because I was breaking out early when there was no ctx or call passed in.

Also fixed reading in the config.  It is read into a map by json and when
going through the map, we screw up the ordering.  We now sort the keys of the
config nodes lexographically so that permissions are correctly inherited by
child nodes.

Change-Id: I6c53a5aa19be6f40415867db540f95d6514e0eed