ref: First cleanup of test/modules, removing RegisterFunction.
This started as a minor CL to clean up the execHandle.envelope
TODO introduced in https://v.io/c/11474, where we noticed that
external commands don't behave correctly.

After speaking with Cos, I've also removed RegisterFunction,
since it's no longer used anywhere.  That lets us remove some of
the interfaces that were presumably added to inject support for
local function calls.

I've also cleaned up the package documentation, so that it shows
up reasonably in godoc.

In addition it would be nice to remove the external command
support completely.  That'll occur separately; it's only used by
the integration test framework, and we may be able to accomplish
that through better factoring of the implementation.

Change-Id: Id70e967a21457f3639106b834d5034c486d96bbc