services/device/device: add glob support to the device tool
The eventual goal is for the device tool to work as follows:
  $ device <command> pattern1 pattern2 ...

Where the patterns are glob patterns that specify a bunch of app instances and
installations. So e.g. device status devmgr/apps/... shows all instances and
installations, or device status devmgr/apps/*/* only shows installations, or
device status devmgr/apps/binaryd/... shows instances and installations for
binaryd.

This cl is the first step: it adds common glob functionality that each command
can then build on (by writing command-specific handlers). We switch the status
command to this new way for illustration, and also add a new command ('ls')
which prints the selected instances and installations to stdout.

More commands to be transitioned to this new scheme in future cls.

Change-Id: I388c4ec039dc19bc16fde76c8889aa437d1ad1f4