"x/ref": Bind identityd macaroon to tool's public key
Presently, the macaroon handed out by identityd as part of the
seekblessings flow is a bearer token, i.e., anyone who posseses
the macaroon can use it to obtain a blessing for the name (email)
encapsulated in the macaroon. This makes the macaroon an attractive
target for theft.

Since macaroons are only meant to be used by the principal tool they
need not be bearer tokens. This CL makes identityd bind the macaroon
to the principal tool's public key (provided by the tool as part of the
"seekblessings" request, see CL: 11169), thereby making it a non-bearer-token
and therefore robust against theft. The public key is subsequently
checked when an RPC is made using the macaroon to obtain a blessing.

Change-Id: I03186c88e4cb9bf128eb0b1ce465fff6eb4821fe