services/device/internal/impl: daemon mode refactoring
This CL restructures the device manager to support execution of
applications in "daemon mode": where the device manager's existing
facilities for monitoring the lifetime of an application can be used
to restart an application on failure.

Restart behaviour is controlled by an implementation of the
RestartPolicy. This CL provides a fixed policy equivalent to the
pre-existing application start functionality.

Change-Id: Ib592bde453054985775a8d40911b3b616c5bb7b0