vdl/compile: Modify CompileExpr to support compilation of package-path qualified identifiers.
CompileExpr used to use a dummy file that didn't have imports that were
compiled into Env.
This caused compilation of package-path quailified identifiers to fail.
We solved this by introducing ExtractExprPackagePaths that allows users to
retrieve all package paths in expressions. Then CompileExpr will include
all needed expr that are also compiled into Env in the file.
Now CompileExprs will succeed in compiling "a/b/c".Foo if "a/b/c" is
in the Env.

Change-Id: I100bf7378b76bb6a4c5b6518774138b31caf94d0