syncbase/server: Implement GlobChildren__ instead of Glob__.
The client-side Glob implementation is now simpler, since it can use the
namespace.Glob method directly.

GlobChildren__ is much simpler than Glob__, and gives us (almost)
exactly the behavior we want.  It can be problematic when a name has
lots of children, because all the children must be returned by
GlobChildren__, and filtering happens at a higher level.  We should
consider passing a pattern to GlobChildren__ so that it can choose to do
its own filtering to save work.

Change-Id: I094b07cc039fcfd9c14d0e92ef2e7ea1223be80f