cmd/principal: Usability tweaks.
Prior to this commit, the command printed by "recvblessings" was
would always fail because of lack of caveats and would also
have to be edited to give a meaningful extension.

This commit changes it so that the exact command can be run successfully
- though it will generate prompts for the user to confirm the lack of
caveats and to add an extension.

For non-interactive uses, the caveat flags and the extension can
still be specified in the command-line.

Change-Id: I0c4fea11883878d776dfcc8a7dd19167e7e73a52