syncbase: store log entries chronologically
This change makes it easier for clients to iterate over log entries
chronologically, which is the common case (both for client watch and for sync
watch). Originally I used reverse chronological order because it was more
efficient for determining the most recent sequence number (done once, at
startup time) - but that's not what we should optimize for.

Change-Id: I84879054ed5617f23f8d717e0e363aee12ea2873