services/device/device: make glob flags per-subcommand instead of global
The main motivation is to customize the default values of the glob flags
depending on the subcommand. For example, update wants to group parallelism by
kind, whereas status is happy to run with full parallelism; also (to come)
update should by default only care about active installations and non-deleted
instances.

Most of the changes are in glob.go, since I tried to package up the flag logic
in one place (to keep code in the specific subcommand files minimal).

Change-Id: I2ec04bd46a3e4ef80f53fe1ea7ef8b3d1ea9032e