services/device/device: introduce test for glob functionality
glob_test.go tests the glob logic directly, instead of relying on the behavior
of various subcommand tests.  To make testing the glob internal library easy,
exported a bunch of identifiers in glob.go (since it's all part of the main
package, there are no practical implications to exporting the
identifiers). While at it, made GlobResults be passed by pointer inside the
library for efficiency.

For now, the only test case verifies that all results appear as expected, and in
the expected order. I'll add more test cases in future cls.

Change-Id: Iea22ddde3db17be3a6e0a6840ef55d54591a9ff7