ref/test/testutil: make use of the rand generator explicit.
Before this change, testutil would implicitly initialize a
random number generator and print out the seed used. This
was mildly annoying because every test binary would print
out this seed as the first thing, but otherwise fine. However,
we'd like to move away from using a global singleton for
logging at which point logging out the seed in an init
function before other intialization has occurred becomes
problematic. This change makes it so that the random number
generator must be initialized explicitly and that this
initialization returns a function to be deferred that prints
out the seed if the test failed.

Change-Id: I0b57ea47d8eaaf45be2a5eeecba2fa1732ae85fd