ref/lib/xrpc:  Add an xserver library that proposes a simplified server API.
The purpose of this API is to reduce boilerplate and make the api easier
to understand.  There are fewer restrictions about which methods can be
called when, and fewer steps to creating a server.

The only flexibility we lose is the ability to listen on multiple ListenSpecs.
Since ListenSpecs already allow users to listen on multiple addresses I expect
this situation will be rare.  When it does come up, users could create
multiple servers with the same service object or dispatcher.

Change-Id: I4ecea023369e16c8cb1408d99ede6ea322cd9ecc