Timestamp, clock api and bugfix.
1) Add timestamp to values committed via watchable store.
2) Introduce a vanadium clock interface which provides
   an estimated UTC timestamp.
3) Fix bug related to sequence number read by Wrap() method
   while creating a watchable store. The bug would end up
   reassigning the sequence number of the last commited tx
   to a new tx to be committed by a newly created wstore
   wrapper object.

This change adds a timestamp field to the LogEntry struct
which is stored for each operation that is committed as
part of a transaction using watchable store. Note that keys
that are not managed by store do not have a LogEntry
associated with them and hence will not have a timestamp.

The clock interface is introduced in this CL but for now
the implementation just uses the system time directly. This
will be addressed in the next CL to come.

Change-Id: I4e0e30ace5efc4388bbc9611d59964ac8ad59cc3