v.io/syncbase/x/ref/services/syncbase/localblobstore/chunker: Introduce data chunker.
Introduce localblobstore/chunker.

Package chunker breaks a stream of bytes into context-defined chunks whose
boundaries are chosen based on content checksums of a window that slides over
the data.

The intent is that when a sequence of bytes is to be transmitted to a recipient
that may have much of the data, the sequnece can be broken down into chunks
based on local context.  The checksums of the resulting chunks may then be
transmitted to the recipient, which can then discover which of the chunks it
has, and which it needs.

Change-Id: I0b1fd8e1a8e5cf31348b9e4120ea46fce9303942