services/device/internal/impl: remove utiltest.V23Init
The only things that the utiltest V23Init wrapper does are:
- unset namespace root env variables
- disable namespace resolution caching

The latter is already part of test.V23Init. The former seems like it should be
present in test.V23Init if at all. But it doesn't seem neecessary since
test.V23Init explicitly sets the namespace roots for the namespace client in the
context. So the env variables could only have an effect on subprocesses if we
pass the env down sloppily. In any case, removing this code doesn't seem to
affect the success of tests, so it seems safe to drop.

Change-Id: I47da3f982d411ddb5d8e34c8f60888a4d266f27b