ref/runtime/internal/flow: Connection sharing flow.Manager API.
This change introduces modifications to the flow.Manager API to allow
connections to be shared between servers and clients on the same Manager.

In particular, Manager.Listen listens for all Acceptors of the manager.
Each Server should call Manager.Acceptor to receive an Acceptor unique
to their server (i.e. the RoutingID of the Acceptor will be unique).
Additional calls to Manager.Dial also automatically Listen on outgoing
connections.

Change-Id: I39b09710c6ca3dedd4a9361a11ff82d066d93f2e