services/internal/servicetest: create shell without a mounttable
test.V23Init already creates a mounttable, and provides the root context with a
deterministic blessing name (test-blessing).  We should use this mounttable, and
the root blessing where possible.  As an example, this CL switches
TestApplicationUpdatePermissions to the simpler scheme.

The change in test/testutil/security.go allows us to conveniently give a default
blessing from one principal to another, and is useful where we don't want to
come up with a brand new principal IDP (as in the example above where we can
just use test-blessing).

Change-Id: I8e6a5e46b6a2937d8415df96008d2fc5b3cf8d4f