services/device/mgmt_v23_test: Use "device publish" and check publisher blessings
Use the "device publish" command instead of binary upload, so that
we publish properly-signed binaries. Check that the device manager
issues a publisher blessing to the signed binary when it's running.

In the process, the test is also made more realistic by using three
principals instead of one -- the device manager is claimed by an
administrator role, the binaries are published by a publisher, and
are run by a user. The device manager also has two blessings now,
representing both a claim from the administrator and a manufacturer
blessing (which is faked for now using the claim).

Using the "device publish" command required changing a few things to
the naming convention that the command expects. This included adding
a known profile for darwin-amd64, which until this point had been
running using "unknown" (!)

Change-Id: I88642cc7117a8db42b5700d463d0062b8d7b3663