syncbase: Add Exists rpc method to app, db, table, row.
Exists returns true if the object exists (has been created).

If the caller doesn't have Read permission on the object, Exists will
return false instead of an error because access errors imply existence
in the current implementation. We leak existence information in other
rpcs, we may want to clean it up.

Change-Id: I7f4300512dff16369932854f0c19e735e3518aa7