syncbase/sb51: Make query demo tool a proper syncbase client.
Replaced query/demo with 'sb51 sh', the first version of a
Syncbase generic client and admin tool.
'sb51 sh' connects to a database on a Syncbase instance,
optionally creating and populating it with demo tables, and
opens a SyncQL shell.
Added abort on error when the shell is not connected to a tty.

Change-Id: I8290df33abdceb49716cb99acd2cb4f25622d84c